util-core: AsyncStream avoid nested function wrapping when parameter-by-name and lazy val used together
Problem

`AsyncSteam.++` uses a lazy val to memorize the
by-name parameter. This function is wrapped into
a new function, which is passed along as the
by-name parameter of the recursive `AsyncSteam.++`
method, and the function wrapping goes on and on.
This creates deeply nested functions as the length
of the stream grows and causes stack overflow.

Solution

Remove lazy val in concat. Use lazy val to define
AsyncStream.Cons.

RB_ID=804408