util-core: Add Var.collectUnsafe that collects without stack overflows
Problem:
Recent changes to Var.collect introduce a stack overflow in Var when
Var.collect is passed a large sequence of Vars.
However, thanks to this recent change, Var.collect correctly orders
updates of Vars, which ensures that the values within Vars being
collected are correct even if some of the collected Vars depend on
others.

Solution:
Introduce Var.collectIndependent, which is the previous version of
Var.collect. This collect does not guarantee the correct ordering
of updates, and thus collected Vars may expose incorrect values if
there are Vars collected that depend on each other. However, this
method does not introduce a potential stack overflow, and is safe
for independent Vars.

Result:
Var.collectIndependent may be used to collect a large number of independent
Vars without risk of a stack overflow.

RB_ID=810514