util-core: Fix leak in Var.apply(T, Event[T]) and Var.patch
Problem

Creating of a Var from an Event via `Var.apply[T](T, Event[T])`
(the canonical way to modify a `Var`) eagerly subscribes to
the passed in `Event`. This subscription keeps a reference to the
resulting `Var` in order to update it. However, because the event
is passed to `Closable.closeOnCollect` it prevents the var from
ever being collected.

The same issue occurs with `Var.patch` as well.

(this great bug report brought to you by Antoine Tollenaere
<atollenaere@twitter.com>)

Solution

Only hold a `WeakReference` back to the `Var` avoiding the strong
reference.

Result

The `Var` can be garbage collected.

RB_ID=809100