Preserve Local settings when logging asynchronously.
Problem

When using QueueingHandler, we are losing the context settings set via Local,
such as TraceId, ClientId, etc., because the Local settings aren't transferred to
the formatting/logging task.

Solution

Use Local.let{} around the formatiing task to let it execute in the context of the
thread that was doing the logging.

Result

When using ansyc logging, the log messages instead of something like this:
  I 0320 14:52:53.600 THREAD46: EchoRequest begin
will now incude full context
  I 0320 14:52:26.006 THREAD46: TraceId:d238dc5aeeb46975 EchoRequest begin

(note, there is also inferClassNames parameter to QueueingHandler, that controls
including of class and method names into the logging messages. This stays unchanged).

Signed-off-by: Kevin Oliver <koliver@twitter.com>

RB_ID=830761