util-core: Reduce allocations in Promise
Problem

`c.t.util.Promise` uses a type-hierarchy, `State[A]` for its internal
state machine. For `Done` and `Linked` this is pure overhead in order
to adapt their internal state to the `State[A]` type.

Solution

Switch to `Any` in order to avoid this overhead. We pay the price of
slightly less maintainable code for a significant reduction in
allocations across all usage of Futures.

Result

Microbenchmarks show the obvious improvement and a production test
showed approximately a 1.8% reduction in allocations. All services
will see a benefit though how much will vary depending on usage
patterns.

RB_ID=832816