util-core: Allow unbounded maxWaiters for AsyncMeter
Problem
-------

Sometimes, the number of potential waiters on a rate limiter are bounded
by some other aspect of the system. In this case, setting the number of
waiters too low can cause spurious failures. The current implementation
of `AsyncMeter` has constant memory overhead for each waiter, which
precludes using a very large value for `maxWaiters`.

Solution
--------

Add a constructor, `newUnboundedMeter`, that uses a linked list-backed
queue rather than an array-backed queue.

Result
------

It is possible to avoid calculating and tuning `maxWaiters` when it is
known to be otherwise bounded.

RB_ID=832060