util-collections: rm apache collections!!!
Problem

We import all of apache commons collections for LruMap, which is a
thin scala LruMap implementation.

Solution

We can replace apache LRUMap with jdk LinkedHashMap.  Cases where
it won't work are probably caches, and we can replace it with either
the underlying apache LRUMap implementation or a guava cache on a
case by case basis.

Result

No more dependency on apache commons collections!

RB_ID=833515
TBR=true