util-zk: Fix race condition in ZkAsyncSemaphore
Problem

Before acquiring a permit, ZkAsyncSemaphore gets the list of existing
permits and then checks each permit ZNode’s data. This raised an error
when one of the existing permits is released between the time the list
was gotten and the data was checked. This was an issue for us because we
are using ZkAsyncSemaphore in Hadoop containers. When many containers
are closing while others are opening, many permits get released rapidly.

Solution

Change the shard permit consensus logic to count lost existing permits
as having not been seen in the first place.

Signed-off-by: Moses Nakamura <mnakamura@twitter.com>

RB_ID=835856