Do not enable scoverage by default
Problem

Enabling scoverage by default causes odd issues with artifacts produced by
`publishLocal`. For example, when running tests	against a project that depends
on artifacts produced by the current util configuration, the following error is
thrown:

    java.io.FileNotFoundException: .../util/util-core/target/scala-2.11/scoverage-data/scoverage.measurements.1147 (No such file or directory)

Solution

Do not set coverage to enabled in Build.scala. Rely on users to manually enable
coverage when it is required (as in travis.yml).