util: Update to cross-compile for scala 2.12
Problem

We want to be able to cross-compile for scala 2.12 but scoverage
isn't ready yet.

Based on this:
https://github.com/twitter/util/commit/efcc8a3d9933886b271e37a59adc19d37722de83

Which was reverted here:
https://github.com/twitter/util/commit/8e7e04d70a0b55f06e0177294ba33bd0d5502b29

Solution

Make scoverage resolve 2.11 artifacts when building for scala
2.12 until 2.12 versions are available.

Result

We should be able to cross-build and this change doesn't appear
to break any downstream projects.

RB_ID=840265