util-core: Memoize should not deadlock reentrant threads
Problem:

In order to ensure its underlying function is only called once for a given
input, Memoize.apply blocks concurrent invocations behind a CountDownLatch until
the first invoc ation completes. Under recursion, it's possible that the same
thread might re-enter Memoize apply with the same input parameters and block on
it's own latch. This effectively converts a bug that would have resulted in a
StackOverflow into a hung thread.

Solution:

Record the thread id of the thread that created the Latch and throw an exception
if the same thread tries to await on that latch.

Result:

Recursive invocations of Memoize's apply with identical input throws an
IllegalStateError.

RB_ID=845702