util-core: Add merge operation to AsyncStream
Problem

Partitions in a pub-sub messaging system map naturally to individual infinite
AsyncStreams. Furthermore, representing the partitions as a single stream and
applying a form of back pressure (using mapConcurrent) on consumption of that
stream and it's partitions is a clear and simple programming model.  Since these
partitions are infinite and no current AsyncStream function matches this usage
it became necessary to invent one.

Solution

A function for merging arbitrary, and possibly infinite, AsyncStreams

Create a function merge, given a sequence of AsyncStreams, that returns the
interleaving of events from it's inputs as they become available.

RB_ID=846681