finagle|util: Make N3 worker pool initialize lazy
Problem

Netty 3 worker pool is always initialized (threads are created, file descriptors registered, etc)
no matter if it's used or not in Finagle (i.e., when Netty 4 is enabled for all clients and servers).

This happens during `GlobalFlag`s initialization, which attempts to load very scala object and check
if it could be a flag. To perform this check (check if a given annotation is present), a class
should be loaded and initialized.

Solution

The RB involves the following changes to make N3 worker pool lazy - only initialised we needed.

  - Ensure `GlobalFlag` API skips Scala package objects that could never be flags
  - Move N3 `WorkerPool` definition from its  own objects into an enclosing `c.t.f.netty3` as `val`

This also moves N4 global flag for the number of workers to make it more user friendly and symmetric
to its N3 version.

Before:

```
 -com.twitter.finagle.netty4$.package$.numWorkers=5
```

After:

```
-com.twitter.finagle.netty4.numWorkers=5
```

RB_ID=865831