util-logging: Correctly increment dropped_records in ScribeHandler
Problem

The dropped_records metric is tracked incorrectly.
The counter is incremented by 1 on each call to incrDroppedRecords(). As a
result, if a batch send of 5 log records fails, the dropped_records metric
will be incremented by 1, but it should be incremented by 5.

Solution

Increment the dropped counter by the 'count' parameter that is passed into
the incrDroppedRecords() method.

RB_ID=866913