util-logging: Allow users to override `c.t.util.logging.Logger` installation
Problem

For users that wish to use more modern logging frameworks
which require the installation of logging bridges, like
SLF4J, the installation of util.app.App's default loggers
causes some intial startup logging to be lost since it
removes/replaces any previously added loggers (like
bridges).

Solution

Move the actual addition of the configured default Logger
Factories to an overridable method. Users can override
this method with an empty implementation to ensure that
any previously configured loggers are retained.

Result

It will be easier for users to work with modern logging
frameworks which required the installation of logging
bridges to bridge as much as the server startup logging
as possible.

RB_ID=870684