util-logging: Pick up log file size
Problem

FileHandler sets 'bytesWrittenToFile' as 0 in 'openLog()'.
This forces log file size to 0 every-time application is restarted, even
if it exists.

Solution

Set 'bytesWrittenToFile' as file size in openStream(), if log file
exists.
Else set it to 0. Remove 'bytesWrittenToFile' from 'openLog()'

Result

As a result of this merge, the FileHandler will rotate log files with
log policy based on Max file size when the limit is reached and prevent
the files from growing too large.

ISSUE: #176

Signed-off-by: Moses Nakamura <mnakamura@twitter.com>

RB_ID=880537