util-stats: Fixed a race condition in CumulativeGauge
Problem

CumulativeGauge had a problem where we could lose gauges when they flapped
between 0 gauges for a name and 1 gauge for a name.

Solution

Java 8 introduces ConcurrentHashMap#computeIfAbsent which simplifies some of the
updating code.  Java 7 introduces Phaser, which allows us to tear down the
gauge permanently after every gauge has declared it wants to be torn down.  Once
we've torn down the gauge, we return a null, signalling to the caller that it
must make a new CumulativeGauge.

Result

CumulativeGauge is now threadsafe, and registration and deregistration are now
asynchronous.

RB_ID=887484