finagle-core: Compute a better logLevel for Failure
Problem

The `logLevel` used in `c.t.f.Failure` factory methods that take a
`Throwable` defaults to using `Level.WARNING`. This is not wrong per
se, but it loses fidelity when the `Throwable` is a
`c.t.logging.HasLogLevel`.

Solution

Compute a more accurate `Level` by traversing the exception chain
looking for `HasLogLevel` and using that when found. If not found, the
existing behavior of using `Level.WARNING` is kept.

RB_ID=894629