util-core: Workaround Scala 2.12 issue in Config.missingValues
Problem

As reported in https://github.com/twitter/util/issues/174 new versions
of Scala 2.12 changed the behavior of what Config.missingValues sees.
Config is deprecated, long in the tooth, a bit finnicky, and generally
unsupported by us.

Solution

Make some patches to the tests to loosen the constraints and also
document that the behavior is not perfect in 2.12.

RB_ID=895199