util-core: Fix memory leak in AsyncStream.mapConcurrent
Problem

There is a subtle memory leak that can occur on long lived streams using
mapConcurrent, where AsyncStream.mapConcurrent$$step will capture a reference to
the outer object. This will cause a reference to the head of the stream to
persist and the stream will accumulate in memory.

Solution

Move the implementation to the companion object the prevent the function from
capturing the reference.

RB_ID=896168