util-core: Memoize the hash code of Buf
Problem / Solution

Computing the hash code of a Buf can be expensive. Buf is immutable, so
we can apply the same strategy found in java.lang.String and memoize the
hash code. Buf is promoted from a trait to an abstract class so that we don't
require scalac to synthesize a field for the cached hash code for each
implementation. This also allows us to remove the java friendly abstract class
AbstractBuf which only existed to turn the trait into an abstract class.

RB_ID=897476