util-core: add Base64Long.fromBase64
Summary:
Problem
-------

util-core supplies functionality for encoding Longs into base-64. This
is useful as a compact serialization of Longs when interacting with
systems that don't support arbitrary binary values (e.g. memcache
keys). util-core does not supply functionality for recovering the Long
from the encoded value, which is useful for debugging.

Solution
--------

Add a method that decodes the base-64 encoded Long value, and
functionality for inverting base-64 alphabet functions.

Test Plan:
Unit test that the decoding is the inverse of the encoding, and that it fails
on out-of-range values. Also expand the existing tests to cover alternate
alphabets.

Reviewers: O5331 source:/util/util-core/, koliver

Reviewed By: O5331 source:/util/util-core/, koliver

Subscribers: koliver

Differential Revision: https://phabricator.twitter.biz/D15674