util-core: Begin replacing ConcatBuf API
Problem

`c.t.io.ConcatBuf` is exposed as a public API which allows users to
create them in ways we'd prefer they did not and it exposes more
implementation than we'd like. For example, users can create an empty
`ConcatBuf` which instead should be a `Buf.Empty`.

Solution

Deprecate `c.t.io.ConcatBuf` and introduce `Buf.apply` to construct
instances from multiple `Bufs`. This also introduces the `Buf.Composite`
trait which implements much of what was done in `ConcatBuf`. There is
a private implementation in `Buf.Composite.Impl` which optimizes `length`
and `isEmpty`.

Result

This allows us to evolve the implementation going forward as seen in
`Buf.Composite.Impl`.

RB_ID=899623