util-core: Begin using Buf.Indexed
Problem

There are is a variety of code utilizing `Bufs` that lend themselves
nicely to efficient random access.

Solution

Begin switching code to utilize the new `Buf.Indexed` APIs.

* Buf.equals
* Buf.hashBuf
* Deadline.tryUnmarshal
* BufReader.remainingUntil
* memcached's ParserUtils.isDigits
* memcached's Framer.extractFrame

Result

The performance improvements in microbenchmarks vary from call to
call. A good high-level signal is that this patch reduces allocations
rates by 1-2% in an example service.

This change also begins work to free us from needing Bufs to be on
heap for alot of operations which is another step towards off-heap
implementations.

RB_ID=905253