finagle-core: Rewrite BufReader in terms of Buf.Indexed
Problem

`BufReader` can be more performant in terms of latency and
allocations.

Solution

Rewrite it in terms of the new `Buf.Indexed` API.

While working on this, bugs in `ByteBufferAsBuf` when `readerIndex` is
not 0 were also addressed.

Result

Significant improvements in terms of latency and allocations in
microbenchmarks.

Using a heap `ByteBufAsBuf`:
- `readByte` 85% faster, 99% less allocations
- `readLongBE` 45% faster, 96% less allocations
- `readBytes` 26% faster, 47% less allocations

Using a direct `ByteBufAsBuf`:
- `readByteDirect` 85% faster, 98% less allocations
- `readLongBE` 77% faster, 96% less allocations
- `readBytes` 31% faster, 96% less allocations

RB_ID=906201