util-core: Add Conversion from Java 8 Optional to Scala Option
Problem

Many Finagle APIs take a Scala `Option` as a parameter. It would
be nice for Java users to be able to work mostly with an `Optional`
instead.

Solution

Add an `asOption` method to the javainterop `Scala` class to convert
from a Java `Optional` to a Scala `Option`.

RB_ID=906512