finagle/finagle-core: Wire Tunables into TimeoutFilter
Problem

We want to be able to configure some modules with Tunables.

Solution

Starting with TimeoutFilter, add a Param constructor that takes
a Tunable and uses it get the timeout for a request.
The constructor is private[twitter] for now, while the API and usage
is still in development.

RB_ID=905395