util/util-core: Make Buf.concat a constant time operation
Problem

`Buf.concat` will make a new collection regardless of the
identity of the two buffers, resulting in O(n) performance
for Composite Bufs, and O(n^2) performance for n concat
operations such as when accumulating data.

Solution

Utilize the O(1) performance of the Vector `:+` and `+:`
operators when we can, resulting in constant time performance
for multiple concat operations. As a bonus, I've also made sure
that a Composite.Impl doesn't contain any Composite.Impl
instances, removing the possibility of a tree structure.

RB_ID=907479