util-tunable: Have StandardTunableMap create a JsonTunableMapper
Problem

StandardTunableMap was using JsonTunableMapper.loadJsonTunables,
which creates a default JsonTunableMapper. We want users to be able
to inject a mapper here in the future.

Solution

In StandardTunableMap, create an instance of a mapper and call that.
This involved some rearrangement of class/object methods
in JsonTunableMapper. We also now pass a path to use, so that in
the future we can pass different paths based on env/instance-id
from ServerInfo.

RB_ID=912079