twitter-server, util-tunable: Endpoint to update Tunables
Problem

We'd like to be able to manipulate in-memory Tunables via
an endpoint in Twitter Server admin.

Solution

Add the /admin/tunables/ endpoint to twitter-server. PUT requests
with JSON bodies can be used to update Tunables for a given id
by making a request to /admin/tunables/$id. Other commands to follow
in separate commits.

RB_ID=912645