twitter-server, util-tunable: Endpoint to clear Tunables
Problem

We'd like to be able to clear in-memory Tunables via
an endpoint in Twitter Server admin.

Solution

Handle DELETE requests to the /admin/tunables/ endpoint on
twitter-server. DELETE requests with JSON bodies can be used to clear
Tunables for a given id by making a request to /admin/tunables/$id.

RB_ID=912969