util-core: All timers now handle negative times/durations uniformly
Problem

Handling negative delays/times in `Timer`s is up to the concrete implementation
and hence is totally different between a number of instances (eg., Netty-based
timer treat negative times/durations as zero while `JavaTimer` throws an exception).
This makes it difficult and almost impossible to substitute one timer with another.

Solution

Enforce the logic of handling negative times/durations in the `Timer` interface
so that all the implementations behave in a similar fashion - treat negative
values as zeros: `Time.epoch` (for time) and `Duration.Zero` (for duration).

Result

It's now possible (and safe) to replace one timer implementation with another
not worrying they will behave differently.

RB_ID=916008