util-core: Introduce Promise.InterruptHandler
Problem

Embedding an interrupt-handler (i.e., a partial function) into a promise is
getting more and more popular within Finagle components given how efficient
it is. This pattern was already encoded in a couple of places and this maybe
a right time to promote it into its own first-class concept.

Solution

Provide a `Promise.InterruptHandler` trait that might be mixed into a `Promise`
to embed/register an interrupt handler.

RB_ID=916750