util-core: Make ByteWriter a true trait
Summary: Problem

The ByteWriter type has two main problems. First, is that it is specialized
for writing to an underlying Array[Byte] which may not  be useful for other
potential implementations, for example, one that writes to an underlying
Netty4 ByteBuf. Second, it has methods that are private[io], meaning it
cannot be extended outside of the com.twitter.io package, thus limiting the
number of useful implementations.

Solution

Make ByteWriter a true trait with only public methods and move the current
definition to a private abstract implementation.

JIRA Issues: CSL-4649

Differential Revision: https://phabricator.twitter.biz/D59996