util, finagle: StatsReceiver cleanups
Summary: # Problem/Solution

This change makes `StatsReceiver.repr` a `def` instead of a `val` so it could be just
delegated through the entire chain of stats receivers (no need to allocate a val in each
wrapping `StatsReceiver`).

In addition to that, this change re-uses `StatsReceiverProxy` where it's possible.

JIRA Issues: CSL-4715

Differential Revision: https://phabricator.twitter.biz/D59762