util-logging: Add a method to Logger to log a throwable that has a log level
Summary: Problem

com.twitter.logging.HasLogLevel allows to mark exceptions with their desired log
level. However, logging these exceptions requires knowledge of the HasLogLevel
trait to extract the level to pass it to the logger object.

Solution

Add a convenience method to com.twitter.logging.Logger - `throwable` - that will
log a throwable with a level that is either a default value or whatever comes
from HasLogLevel.

Differential Revision: https://phabricator.twitter.biz/D69136