util-security: Assert validity of X.509 certificates when read from a file
Summary: Problem

We do not currently assert the validity of a well-formed X.509 certificate
when reading the contents from a file. That is, no comparison of the current
date in relation to the valid range specified in the file is performed.

Solution

Perform a validity check when reading the contents of an X.509 certificate
file. Invalid date ranges will result in either a CertificateExpiredException
or a CertificateNotYetValidException, depending.

JIRA Issues: CSL-5125

Differential Revision: https://phabricator.twitter.biz/D88745