util, finagle, finatra: Fix publishing of root project
Summary: Problem

Finatra was still trying to publish its root module. This was due to
the `publishTo` key not being set. Interestingly, util and finagle
were working, but it appears to be due to ordering and how that key
was set.

Solution

Explictly set the `publishTo` key to a dummy value.

Result

Finatra will no longer try to publish its root project and
util/finagle's build files are more robust to this regressing in the
future.

JIRA Issues: csl-5171

Differential Revision: https://phabricator.twitter.biz/D95145