util/finagle: Rewrite `Future.onSuccess` as `Future.respond`
Summary: Problem

`Future.onSuccess` is written in terms of `Future.respond` and code
that calls `onSuccess` then has an allocation of an additional
`Function1` to adapt to the `respond` API.

Solution

Use an experimental scalafix [1] rewrite rule to do the initial
rewrite from `onSuccess` to `respond`.

Result

Less allocations in util and finagle.

[1] https://github.com/scalacenter/scalafix

JIRA Issues: csl-5258

Differential Revision: https://phabricator.twitter.biz/D95553