util-core: Fix buf in Buf.ByteArray.Shared.apply(bytes, begin, end)
Summary: Problem

We are incorrectly using the j.u.Arrays.copyOfRange function, providing
its 'to' argument with a 'end - begin'. This has two problems: first, if
begin isn't 0, we're going to truncate. Second, if end > data.length,
we're going to include a bunch of bonus 0's at the end of the buffer. As
a bonus, we don't validate the begin and end parameters in the Owned and
Shared apply methods, meaning we can get situations such as
`Buf.ByteArray.Owned(data, 5, 0).length == -5`.

Solution

Make the copy correctly and validate our input using the same rules as
the `Buf.slice` method.

JIRA Issues: CSL-5354

Differential Revision: https://phabricator.twitter.biz/D100648