Fix strict_deps for a large number of targets
Summary: This is a semi-automated change. We're not expecting you to review the individual changes to the BUILD files you own, but we want you to be aware that they are happening.

A bug in Pants (DPB-9305) means these targets were not properly strict_deps-ified. This adds those missing dependencies.

Strict deps restricts what is on the compile classpath to improve compile times. This means you spend less time waiting for your builds. However, with this change, you need to remember that transitive dependencies don't work so if you need something to compile code, you need to add it directly to your dependencies.

Thanks to @stuhood for putting together a large number of these fixes!

JIRA Issues: DPB-9305

Differential Revision: https://phabricator.twitter.biz/D104958