util: TravisCI config runs utilIntelliJ project multiple times
Summary: Problem/Solution

The utilIntelliJ project should only run under a Scala version
of 2.11.x. However, we have a matrix of versions configured in
the .travis.yml -- 2.11.11 and 2.12.4. Everything in the script
section gets run per scala version. The script statement for
utilIntelliJ is hardcoded to 2.11.11 but this means it builds
twice, each time with the hardcoded scala version.

Update the script section to only run the utilIntelliJ script
if the scala version starts with "2.11".

Differential Revision: https://phabricator.twitter.biz/D119673