util-core: Implicit conversion from "x.percent" to fractional Double
Summary: Problem

It would be useful to have a conversion between "x.percent"
and a fractional Double so that argument passing and usage
example of functions/constructors taking a fraction-as-percentage
argument are more explicit. For example, the user is clear that 50.percent
means 50%, but 0.5 could be 0.5% or 50% and necessitates
reading the documentation.

Solution

Add an implicit "x.percent" conversion to a fractional Double.

JIRA Issues: CSL-5835

Differential Revision: https://phabricator.twitter.biz/D128792