util-app: Ensure flag parsing error reason is emitted before printing Flags#usage
Summary: Problem

Currently when flag parsing fails we attempt to concatenate the reason for the
failure (i.e., which flag input failed to be parsed) with the result of
`Flags#usage` which collects and prints all available flags to be set. The
issue is that collecting the flag usage can be an expensive operation in that
the code attempts to scan the entire classpath for all global flags. This can
fail in some situations (with an OOM) due to the application having a fairly
large classpath.

Solution

We want to ensure that users get the error reason before attempting to collect
and print all available flags of the system, thus we break out calling
`Flags#usage` from the error reason and delay the invocation until after we
have printed (and flushed) the error reason to System.err.

Note:

These changes do not attempt to solve the problem of an application running into
an OOM when `Flag#usage` fails due to a large classpath. It only attempts to
actually provide the useful error reason regardless of the OOM happening.

JIRA Issues: CSL-5597

TBR=true

Differential Revision: https://phabricator.twitter.biz/D137629