util-logging: Fix incorrect `loggerName` in `c.t.logging.ScribeHandler`
Summary: Problem

The `c.t.logging.ScribeHandler` tries to short-circuit publishing of
messages that come from itself by doing a comparison on the loggerName
in the record to be published against it's own captured `loggerName`
representing the name of the Logger configured for the `ScribeHandler`.

Unfortunately, the `ScribeHandler` uses `getClass.toString` to set it's
`loggerName`. While the Logger is constructed from the class by calling
`getClass.getName`.

The Java implementation for `Class#toString` prepends "interface" or
"class" to the class name in the returned String. This means that comparing
`getClass.toString` to `getClass.getName` will never match and thus the logic
for short-circuiting in the `ScribeHandler` is broken.

Solution

Update the `ScribeHandler#loggerName` to be `getClass.getName` which will
match how the Logger is named.

Result

The `ScribeHandler` will correctly skip trying to publish log lines emitted
from itself as intended.

JIRA Issues: CSL-3685

Differential Revision: https://phabricator.twitter.biz/D161552